break up Manipulation hierarchy
